@@ -2,6 +2,6 @@ module.exports = {
   routes: { test: './test/routeIndex' },
   server: 'https://api.choerodon.com.cn',
   dashboard: {
-    'iam': 'test/dashboard/*',
+    iam: 'test/dashboard/*',
   },
 };
