@@ -1,13 +1,14 @@
+import path from 'path';
+import mkdirp from 'mkdirp';
+
 let isInitialized = false;
 exports.initialize = function (context) {
   if (isInitialized) {
     console.error('`context` had been initialized');
     return;
   }
+  const tmpDirPath = context.tmpDirPath = path.join(__dirname, '../../../tmp');
+  mkdirp.sync(tmpDirPath);
   Object.assign(exports, context);
   isInitialized = true;
-};
-
-exports.turnOnSSRFlag = function () {
-  exports.isSSR = true;
 };
