@@ -6,7 +6,6 @@ import Card from './Card';
 @inject('DashboardStore')
 @observer
 export default class Column extends Component {
-
   static defaultProps = {
     column: [],
   };
@@ -71,9 +70,12 @@ export default class Column extends Component {
 
   render() {
     const { column, DashboardStore } = this.props;
-    if (column.length > 0 && (DashboardStore.editing || column.filter(({ visible }) => visible).length > 0)) {
+    if (column.length > 0
+      && (DashboardStore.editing || column.filter(({ visible }) => visible).length > 0)) {
       return (
-        <div>{column.map(item => this.renderCard(item))}</div>
+        <div>
+          {column.map(item => this.renderCard(item))}
+        </div>
       );
     } else {
       const { prefixCls, dragData } = this.props;
