@@ -4,7 +4,7 @@ import { withRouter } from 'react-router-dom';
 import classNames from 'classnames';
 import { Button, Icon } from 'choerodon-ui';
 import findFirstLeafMenu from '../util/findFirstLeafMenu';
-import { historyPushMenu, getMessage } from '../../common';
+import { getMessage, historyPushMenu } from '../../common';
 
 @withRouter
 @inject('AppState', 'MenuStore')
@@ -12,7 +12,7 @@ import { historyPushMenu, getMessage } from '../../common';
 export default class Setting extends Component {
   getGlobalMenuData = () => {
     const { MenuStore, history } = this.props;
-    MenuStore.loadMenuData({ type: 'site' }, false).then(menus => {
+    MenuStore.loadMenuData({ type: 'site' }, false).then((menus) => {
       if (menus.length) {
         const { route, domain } = findFirstLeafMenu(menus[0]);
         historyPushMenu(history, route, domain);
@@ -28,7 +28,7 @@ export default class Setting extends Component {
     return (
       <Button className={classString} onClick={this.getGlobalMenuData}>
         {getMessage('管理', 'Manage')}
-        <Icon className="manager-icon" type="settings " style={{ marginLeft: '5px' }}/>
+        <Icon className="manager-icon" type="settings " style={{ marginLeft: '5px' }} />
       </Button>
     );
   }
