@@ -2,6 +2,6 @@ import React, { Component } from 'react';
 
 export default class Document extends Component {
   render() {
-    return `文档`;
+    return '文档';
   }
 }
