@@ -2,6 +2,6 @@ import React, { Component } from 'react';
 
 export default class Guide extends Component {
   render() {
-    return `新手指引`;
+    return '新手指引';
   }
 }
